refactor: consolidate qwen/grok environments into shared OpenAI-compatible builder
Extract common environment setup patterns from qwen and grok providers
into a reusable `buildOpenAICompatibleEnvironment()` function.

Before: qwen (170 lines) + grok (146 lines) = 316 lines
After: shared builder (195 lines) + qwen (35 lines) + grok (23 lines) = 253 lines
Net reduction: 63 lines (-20%)

The shared builder handles:
- Config directory creation and telemetry cleanup
- Standard lifecycle hooks (session start, stop, completion)
- Settings.json with selectedAuthType = "openai"
- Memory protocol integration
- GitHub Projects context injection
- Instructions file creation
- Shell wrappers and provider override handling

This is the first step in "provider adapter narrowing" from the 60-day roadmap.
