+++ packages/shared/src/providers/qwen/environment.ts
@@ -2,168 +2,40 @@ import type {
   EnvironmentContext,
   EnvironmentResult,
 } from "../common/environment-result";
-import {
-  getMemoryStartupCommand,
-  getMemorySeedFiles,
-  getProjectContextFile,
-} from "../../agent-memory-protocol";
-import { buildGenericInstructionsContent } from "../../agent-instruction-pack";
-import { getTaskSandboxWrapperFiles } from "../common/task-sandbox-wrappers";
-import { buildStandardLifecycleHooks } from "../../provider-lifecycle-adapter";
+import { buildOpenAICompatibleEnvironment } from "../common/openai-compatible-environment";
 
-// Prepare Qwen CLI environment for OpenAI-compatible API key mode.
-// We previously supported the Qwen OAuth device flow, but cmux now uses
-// API keys via DashScope or OpenRouter configured in Settings.
-async function makeQwenEnvironment(
-  ctx: EnvironmentContext,
-  defaultBaseUrl: string | null,
-  defaultModel: string | null
-): Promise<EnvironmentResult> {
-  const { Buffer } = await import("node:buffer");
-
-  // useHostConfig is safe for desktop/Electron apps where the host IS the user's machine.
-  // For server deployments, this should be false to prevent credential leakage.
-  const useHostConfig = ctx.useHostConfig ?? false;
-
-  const files: EnvironmentResult["files"] = [];
-  const env: Record<string, string> = {};
-  const startupCommands: string[] = [];
-
-  // Ensure .qwen directory exists
-  startupCommands.push("mkdir -p ~/.qwen");
-
-  // Clean up any old Qwen telemetry files from previous runs
-  // The actual telemetry path will be set by the agent spawner with the task ID
-  startupCommands.push("rm -f /tmp/qwen-telemetry-*.log 2>/dev/null || true");
-
-  // Build standard lifecycle hooks using shared adapter
-  const lifecycleHooks = buildStandardLifecycleHooks(
-    {
-      provider: "qwen",
-      taskRunId: ctx.taskRunId,
-      includeMemorySync: true,
-      createCompletionMarker: true,
-    },
-    Buffer.from.bind(Buffer)
-  );
-  files.push(...lifecycleHooks.files);
-  startupCommands.push(...lifecycleHooks.startupCommands);
-
-  // Fire session start hook on sandbox initialization
-  startupCommands.push("/root/lifecycle/qwen/session-start-hook.sh &");
-
-  type QwenSettings = {
-    selectedAuthType?: string;
-    useExternalAuth?: boolean;
-    [key: string]: unknown;
-  };
-
-  let settings: QwenSettings = {};
-  if (useHostConfig) {
-    const { readFile } = await import("node:fs/promises");
-    const { homedir } = await import("node:os");
-    const { join } = await import("node:path");
-    // Merge/update ~/.qwen/settings.json with selectedAuthType = "openai"
-    const qwenDir = join(homedir(), ".qwen");
-    const settingsPath = join(qwenDir, "settings.json");
-    try {
-      const content = await readFile(settingsPath, "utf-8");
-      try {
-        const parsed = JSON.parse(content) as unknown;
-        if (parsed && typeof parsed === "object") {
-          settings = parsed as QwenSettings;
-        }
-      } catch {
-        // Ignore invalid JSON and recreate with defaults
-      }
-    } catch {
-      // File might not exist; we'll create it
-    }
-  }
-
-  // Force OpenAI-compatible auth so the CLI doesn't ask interactively
-  settings.selectedAuthType = "openai";
-  // Ensure we don't try an external OAuth flow in ephemeral sandboxes
-  if (settings.useExternalAuth === undefined) {
-    settings.useExternalAuth = false;
-  }
-
-  const mergedContent = JSON.stringify(settings, null, 2) + "\n";
-  files.push({
-    destinationPath: "$HOME/.qwen/settings.json",
-    contentBase64: Buffer.from(mergedContent).toString("base64"),
-    mode: "644",
-  });
-
-  // Set sensible default base URL for the OpenAI-compatible API if none provided via settings
-  if (defaultBaseUrl) env.OPENAI_BASE_URL = defaultBaseUrl;
-  if (defaultModel) env.OPENAI_MODEL = defaultModel;
-
-  // Provider override takes precedence over default base URL
-  if (ctx.providerConfig?.isOverridden && ctx.providerConfig.baseUrl) {
-    env.OPENAI_BASE_URL = ctx.providerConfig.baseUrl;
-  }
-
-  // Add agent memory protocol support
-  startupCommands.push(getMemoryStartupCommand());
-  files.push(...getMemorySeedFiles(ctx.taskRunId, ctx.previousKnowledge, ctx.previousMailbox, ctx.orchestrationOptions, ctx.previousBehavior));
-
-  // Inject GitHub Projects context if task is linked to a project item (Phase 5)
-  if (ctx.githubProjectContext) {
-    files.push(
-      getProjectContextFile({
-        ...ctx.githubProjectContext,
-        taskRunJwt: ctx.taskRunJwt,
-        callbackUrl: ctx.callbackUrl,
-      }),
-    );
-  }
-
-  // Add QWEN.md with memory protocol instructions for the project
-  const qwenMdContent = buildGenericInstructionsContent({
-    policyRules: ctx.policyRules,
-    orchestrationRules: ctx.orchestrationRules,
-    previousBehavior: ctx.previousBehavior,
-    isOrchestrationHead: ctx.isOrchestrationHead,
-  }, "# cmux Project Instructions");
-  files.push({
-    destinationPath: "/root/workspace/QWEN.md",
-    contentBase64: Buffer.from(qwenMdContent).toString("base64"),
-    mode: "644",
-  });
-
-  // Block dangerous commands in task sandboxes (when enabled via settings)
-  // Disabled by default - use permission deny rules or policy rules instead
-  const hasTaskRunJwt = ctx.taskRunJwt.trim().length > 0;
-  if (hasTaskRunJwt && ctx.enableShellWrappers) {
-    files.push(...getTaskSandboxWrapperFiles(Buffer));
-  }
-
-  return { files, env, startupCommands };
-}
-
-// OpenAI-compatible mode without provider defaults.
-// Base URL and model are supplied via env (Settings):
-//  - DashScope: set OPENAI_API_KEY and (optionally) OPENAI_BASE_URL + OPENAI_MODEL
-//  - OpenRouter: set OPENROUTER_API_KEY (server maps to OPENAI_API_KEY) and optional OPENAI_MODEL
+/**
+ * Qwen CLI environment for OpenRouter.
+ *
+ * Uses OpenAI-compatible API with OpenRouter endpoint.
+ * API key is provided via OPENROUTER_API_KEY env var (server maps to OPENAI_API_KEY).
+ */
 export async function getQwenOpenRouterEnvironment(
   ctx: EnvironmentContext
 ): Promise<EnvironmentResult> {
-  // Hardcode OpenRouter compatible endpoint and default Qwen model.
-  return makeQwenEnvironment(
-    ctx,
-    "https://openrouter.ai/api/v1",
-    "qwen/qwen3-coder:free"
-  );
+  return buildOpenAICompatibleEnvironment(ctx, {
+    provider: "qwen",
+    configDir: "qwen",
+    instructionsPath: "/root/workspace/QWEN.md",
+    defaultBaseUrl: "https://openrouter.ai/api/v1",
+    defaultModel: "qwen/qwen3-coder:free",
+  });
 }
 
+/**
+ * Qwen CLI environment for DashScope/ModelStudio.
+ *
+ * Uses OpenAI-compatible API with DashScope International endpoint.
+ * API key is provided via DASHSCOPE_API_KEY env var (server maps to OPENAI_API_KEY).
+ */
 export async function getQwenModelStudioEnvironment(
   ctx: EnvironmentContext
 ): Promise<EnvironmentResult> {
-  // Hardcode DashScope Intl (ModelStudio) endpoint and qwen3-coder-plus model.
-  return makeQwenEnvironment(
-    ctx,
-    "https://dashscope-intl.aliyuncs.com/compatible-mode/v1",
-    "qwen3-coder-plus"
-  );
+  return buildOpenAICompatibleEnvironment(ctx, {
+    provider: "qwen",
+    configDir: "qwen",
+    instructionsPath: "/root/workspace/QWEN.md",
+    defaultBaseUrl: "https://dashscope-intl.aliyuncs.com/compatible-mode/v1",
+    defaultModel: "qwen3-coder-plus",
+  });
 }
